# contracts/openapi/iac.v1.yaml
...
openapi: 3.1.0
info:
  title: Nebula IaC Service
  version: "1.0.0"
  summary: Run individual terraform operations as asynchronous jobs against a checked-out workspace.
  description: |
    The IaC service is a raw executor for the Terraform CLI. Each POST
    endpoint enqueues a **job** that runs exactly one terraform command
    (`init`, `validate`, `plan`, `show`, `apply`, or `import`) against
    the workspace at `workspace_path` and returns `202 Accepted` with a
    job id immediately. Clients fetch progress and results by polling
    `GET /v1/jobs/{job_id}`.

    Orchestration is the caller's job: the service does not sequence
    commands, interpret plan output, or detect drift. A caller that
    wants "validate this workspace" submits `init`, then `validate`,
    then `plan` (and `show` if it needs the plan JSON) as separate
    jobs, deciding after each one whether to continue.

    Jobs targeting the same `workspace_path` execute one at a time in
    submission (FIFO) order; jobs for different workspaces may run
    concurrently. A submission is never rejected because the
    workspace is busy — it queues.

    Two failure planes are kept distinct:

    * **Terraform-level failures** are normal outcomes: the job ends
      `succeeded` and its `OperationResult` carries the command's
      `exit_code` (non-zero on failure) plus the raw `stdout` and
      `stderr`. The service passes terraform's output through
      verbatim — including provider authentication errors.
...
servers:
...
paths:
...
  /v1/validate:
    post:
      operationId: validate
      tags: [validate]
      summary: Enqueue a `terraform validate` job for a workspace.
      description: |
        Enqueues a job that runs `terraform validate` against the
        workspace at `workspace_path`, then returns `202 Accepted`
        immediately. Poll `GET /v1/jobs/{job_id}` for the
        OperationResult carrying the command's exit code and raw
        output. The workspace must already be initialised (submit an
        `init` job first).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
      responses:
        "202":
          description: |
            Job accepted and queued. Jobs for the same `workspace_path`
            run one at a time in submission order; poll
            `GET /v1/jobs/{job_id}` for progress and the result.
          headers:
            Location:
              description: URL of the created job resource.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "400":
          description: |
            Request could not be parsed at the transport layer. The OSS
            reference impl returns the framework-default body for this
            case (see `application/json`).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Token is valid but not authorized to call this endpoint.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: |
            `workspace_path` does not exist or is not accessible from
            the implementation. Distinct from a terraform-level
            failure, which surfaces in the job result.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Request body failed schema validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: |
            Implementation is running but missing prerequisites needed
            to accept jobs (terraform binary, configured state backend,
            etc.). Faults that only surface while the job runs end the
            job as `failed` instead.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /v1/apply:
    post:
      operationId: apply
      tags: [apply]
      summary: Enqueue a `terraform apply` job for a plan file.
      description: |
        Enqueues a job that runs `terraform apply <plan_file>` against
        the workspace at `workspace_path`, then returns `202 Accepted`
        immediately. Poll `GET /v1/jobs/{job_id}` for the
        OperationResult carrying the command's exit code and raw
        output. The `plan_file` must have been written by a previous
        `plan` job on the same workspace; applying is exactly the
        changes recorded in that file.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRequest"
      responses:
        "202":
          description: |
            Job accepted and queued. Jobs for the same `workspace_path`
            run one at a time in submission order; poll
            `GET /v1/jobs/{job_id}` for progress and the result.
          headers:
            Location:
              description: URL of the created job resource.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "400":
          description: |
            Request could not be parsed at the transport layer. The OSS
            reference impl returns the framework-default body for this
            case (see `application/json`).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Token is valid but not authorized to call this endpoint.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: |
            `workspace_path` does not exist or is not accessible from
            the implementation. Distinct from a terraform-level
            failure, which surfaces in the job result.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Request body failed schema validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: |
            Implementation is running but missing prerequisites needed
            to accept jobs (terraform binary, configured state backend,
            etc.). Faults that only surface while the job runs end the
            job as `failed` instead.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Per-service bearer token. The token value is the contents of the
        environment variable named in the system config (default
        `NEBULA_IAC_TOKEN`). Implementations MAY accept all
        requests when no token is configured for local development.
  schemas:
...
    ValidateRequest:
      type: object
      required: [workspace_path]
      additionalProperties: false
      properties:
        workspace_path:
          type: string
          description: |
            Absolute filesystem path to the Terraform workspace as visible
            to the implementation. For the OSS docker-compose deployment,
            this is a path on the shared volume mounted into both the
            core and the IaC service.
          minLength: 1
          maxLength: 4096
          examples: ["/workspaces/my-iac_a1b2c3"]
        scope_id:
          type: [string, "null"]
          description: |
            Cloud provider scope the operation targets — Azure:
            subscription id, GCP: project id, AWS: account id. Used as a
            fallback when the credentials resolved for the workspace do
            not already carry a scope. Implementations MAY ignore it when
            the resolved credentials are fully specified.
          maxLength: 1024
...
    ApplyRequest:
      type: object
      required: [workspace_path, plan_file]
      additionalProperties: false
      properties:
        workspace_path:
          type: string
          description: |
            Absolute filesystem path to the Terraform workspace as visible
            to the implementation. For the OSS docker-compose deployment,
            this is a path on the shared volume mounted into both the
            core and the IaC service.
          minLength: 1
          maxLength: 4096
          examples: ["/workspaces/my-iac_a1b2c3"]
        scope_id:
          type: [string, "null"]
          description: |
            Cloud provider scope the operation targets — Azure:
            subscription id, GCP: project id, AWS: account id. Used as a
            fallback when the credentials resolved for the workspace do
            not already carry a scope. Implementations MAY ignore it when
            the resolved credentials are fully specified.
          maxLength: 1024
        plan_file:
          type: string
          description: |
            Filename (not a path) of a plan file previously written by a
            `plan` job on this workspace; `apply` executes exactly the
            changes recorded in it. Restricted to a single path segment
            so it cannot escape the workspace.
          pattern: "^[A-Za-z0-9._-]{1,128}$"
          examples: ["a1b2c3d4e5f6.plan"]
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed]
      description: |
        `queued`: accepted, waiting for its workspace's FIFO queue.
        `running`: terraform command executing.
        `succeeded`: command ran to completion — inspect `result` for
        the terraform-level outcome (`exit_code` may be non-zero).
        `failed`: service-level fault — inspect `error`.
...
    JobAccepted:
      type: object
      description: Returned by the submitting endpoints when a job is enqueued.
      required: [job_id, status]
      additionalProperties: false
      properties:
        job_id:
          type: string
          format: uuid
          description: Identifier to poll at `GET /v1/jobs/{job_id}`.
        status:
          $ref: "#/components/schemas/JobStatus"
          description: Always `queued` at submission time.
...
    Problem:

      type: object
      description: |
        Problem details for HTTP APIs (RFC 7807).

      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: "about:blank"
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference